# backend/openapi.yml
openapi: 3.0.3
info:
  title: STGY API
  version: 1.0.0
  description: |
    STGY SNS backend API (Express/TypeScript)
    - Cookie authentication: session_id
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    PostDetail:
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            content: { type: string }
          required: [content]
...
    PostLite:
      type: object
      properties:
        id: { type: string }
        ownedBy: { type: string }
        replyTo: { type: string, nullable: true }
        allowLikes: { type: boolean }
        allowReplies: { type: boolean }
        createdAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        ownerNickname: { type: string }
        ownerLocale: { type: string }
        replyToOwnerId: { type: string, nullable: true }
        replyToOwnerNickname: { type: string, nullable: true }
        countLikes: { type: integer }
        countReplies: { type: integer }
        tags:
          type: array
          items: { type: string }
      required:
        [
          id,
          ownedBy,
          replyTo,
          allowLikes,
          allowReplies,
          createdAt,
          publishedAt,
          updatedAt,
          ownerNickname,
          ownerLocale,
          replyToOwnerId,
          replyToOwnerNickname,
          countLikes,
          countReplies,
          tags,
        ]
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /posts/search:
    get:
      summary: Search posts (Full text search)
      security:
        - sessionCookie: []
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 21 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostLite"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /posts/count:
    get:
      summary: (Admin only) Get post count
      security:
        - sessionCookie: []
      parameters:
        - name: query
          in: query
          schema: { type: string }
        - name: ownedBy
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: replyTo
          in: query
          schema: { type: string }
          description: |
            If omitted: include all posts.
            If provided as empty string: treated as null (non-reply posts).
            Otherwise: treated as an exact replyTo id.
      responses:
        "200":
          description: Post count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                required: [count]
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /posts:
    get:
      summary: List posts
      security:
        - sessionCookie: []
      parameters:
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: after
          in: query
          schema:
            type: string
            pattern: '^(?:0x)?[0-9a-fA-F]{1,16}$'
          description: Exclusive hexadecimal post ID cursor. Supported with order=asc or order=desc and offset=0. Ascending order returns IDs greater than after; descending order returns IDs less than after.
        - name: query
          in: query
          schema: { type: string }
          description: Admin-only PostgreSQL substring search filter. Use /posts/search for normal full-text search.
        - name: ownedBy
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: replyTo
          in: query
          schema: { type: string }
          description: |
            If omitted: include all posts.
            If provided as empty string: treated as null (non-reply posts).
            Otherwise: treated as an exact replyTo id.
        - name: focusUserId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"

    post:
      summary: Create post
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string, description: Admin only; ignored/forbidden for non-admin. }
                content: { type: string }
                locale: { type: string, nullable: true }
                ownedBy: { type: string, description: Admin only; defaults to current user. }
                replyTo: { type: string, nullable: true }
                publishedAt: { type: string, nullable: true }
                allowLikes: { type: boolean }
                allowReplies: { type: boolean }
                tags:
                  type: array
                  items: { type: string }
              required: [content, tags]
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /posts/by-followees:
    get:
      summary: List posts by followees
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: includeSelf
          in: query
          schema: { type: boolean, default: false }
        - name: includeReplies
          in: query
          schema: { type: boolean, default: true }
        - name: focusUserId
          in: query
          schema: { type: string }
        - name: limitPerUser
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Posts list by followees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /posts/liked:
    get:
      summary: List posts liked by user
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: after
          in: query
          description: Exclusive liked post ID cursor; requires offset=0
          schema: { type: string }
        - name: includeReplies
          in: query
          schema: { type: boolean, default: true }
        - name: focusUserId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Posts liked by user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
  /posts/{id}/lite:
    get:
      summary: Get post lite by id
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Post lite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLite"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
  /posts/{id}:
    get:
      summary: Get post by id
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: focusUserId
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"

    put:
      summary: Update post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
                locale: { type: string, nullable: true }
                ownedBy: { type: string, description: Admin only; forbidden for non-admin. }
                replyTo: { type: string, nullable: true }
                publishedAt: { type: string, nullable: true }
                allowLikes: { type: boolean }
                allowReplies: { type: boolean }
                tags:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "400":
          $ref: "#/components/responses/Error400"

    delete:
      summary: Delete post
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { type: string }
                required: [result]
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
...
  /posts/pub/{id}:
    get:
      summary: Get a public post by id (published)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PubPostDetail"
        "400":
          $ref: "#/components/responses/Error400"
        "404":
          $ref: "#/components/responses/Error404"
...